feat(accent): support shifted accented characters
Thanks to @sigprof, I was able to implement a working change to get
lowercased and uppercased accented characters.

- On leader-key trigger, the current mod is checked and stored in a global
  variable.
- During the leader sequence checks, if the leader_shifted is on, we
  tap_code16 with S().

This means that for an accented character, the _leader_ key must be
shifted, not any of the leader sequence keys. This changes slightly from
how Espanso was doing it (e.g. `,Eg` now has to be `<S+Leader>eg`) but
it's as good.